ml-api/requirements.txt
uvicorn[standard]==0.13.4
gunicorn==20.1.0
fastapi==0.63.0
python-multipart==0.0.5
psycopg2-binary==2.8.6
configparser==5.0.2
passlib==1.7.4
python-jose==3.2.0
pycryptodome==3.10.1
alembic==1.6.5
datetime==4.3
celery==5.1.2
flower==1.0.0
redis==3.5.3
numpy==1.21.2
pillow==8.3.2
opencv-python==4.5.3.56